This commit removes redirection of stderr of executed commands to stdout
This causes a bug in DRD after recent modifications in dune-client.
For instance, a warning (syntax warning: for bar) has been added to
dune-client's stderr. This caused a parsing issue in DRD when retrieving
the list of known contracts and trying to parse it because of the
expected pattern <alias>: <pkh>.

But bug is simply fixed by ignoring stderr (which should always be done)